Moved all ConverterFactory classes into the same header file
